Delay for the minimum CSN time after SPI transfer
The minimum chip select high time is 2*tCLK + 10 ns. This solution waits for this amount of time after each SPI transaction. It accounts for the clock speed of the arduino using F_CPU, and the clock speed of the TMC5160 with TMC5160::fclk.
This delay is only necessary when performing consecutive SPI transactions. But, it's so short (~200 ns) that there's no real cost to adding the delay after every SPI transaction.